pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "xcrape"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.10",
    "httpx>=0.27",
    "tenacity>=8.2",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "testcontainers[postgresql]>=4.0",
    "ruff>=0.4",
    "black>=24.0",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/xcrape"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
strict = true
python_version = "3.12"
mypy_path = "src"
